SMAX: Reorganized GetComponentsResponseDto
