+++ Templates/Devon4NetAPI/Devon4Net.WebAPI.Implementation/Business/SmaxHcmrManagement/Controllers/SmaxHcmController.cs
@@ -6,7 +6,6 @@ using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.CreateDesignVersion;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.ServiceDesigner;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.ServiceDesigner.ApplyComponentTemplateToComponent;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.ServiceDesigner.CreateComponentsAndRelations;
-using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.ServiceDesigner.GetNodeProperties;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Offering;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Providers;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Request.CreateRequest;
